[PX] +: Create new module called Init
Based on the original Kega Init module.
The module will enable me to easaly create config options like category ID entries used for the store info which is based on categories.
This version of the module is giving an 404 error page though but I have not figured out yet why.
Story continues...