[PX] +: implement animated product list items
In preparation I already created plain html to start working creating the animation for the products on the listpage.
In order to get this to work I had to load a jquery lib and create a new css file for customizations
The animations and interactions are not working as they should but its a start!