[PX] +: optimize listpage template & improve js script
I was able to use the existing classnames in the store_list template. Therefor the wrapper classname are no longer necessary. This also enhances the flexibility for implementation on other pages or projects.
I also found some flaws in the js script causing the interactions to malfunction. I was able to fix a few things but not everything yet.
Missing part is collaps the expanded tab when other tab is clicked.