[PX] +: setup store info structure on listpage
I had to use the small_image twice because the base image is not support by default on the listpage.
Without haveing to adjust this in the database I decided to go this way.
It is possible to change this in the database and set listpage support to true but this was not the most relevant part of this exersice and therefor not important to change at this stage of building.

I've create two content div's because this makes it easy to set an overflow hidden which is used to hide exessive content.